// examples/boilerplate/src/types.ts
export interface FileType {
  id: string;
  name: string;
  size: number;
  type: string;
  url: string;
}